fix: resolve clippy warnings and build errors
- Fix missing Value import in dew-quaternion lua.rs (cfg-gated)
- Consolidate trait bounds to avoid multiple_bound_locations warning
- Apply clippy auto-fixes for collapsible_if and unnecessary_clone
- Simplify conditional logic in optimizer passes

All crates now build cleanly with --all-features.

+++ crates/dew-linalg/src/optimize.rs
@@ -125,7 +125,7 @@ impl ConstValue {
     /// Try to interpret an AST as a constant value.
     fn from_ast(ast: &Ast) -> Option<Self> {
         match ast {
-            Ast::Num(n) => Some(ConstValue::Scalar(*n as f64)),
+            Ast::Num(n) => Some(ConstValue::Scalar(*n)),
 
             Ast::Call(name, args) => match (name.as_str(), args.len()) {
                 ("vec2", 2) => {
@@ -158,7 +158,7 @@ impl ConstValue {
     /// Helper to extract a scalar from an AST.
     fn as_scalar(ast: &Ast) -> Option<f64> {
         match ast {
-            Ast::Num(n) => Some(*n as f64),
+            Ast::Num(n) => Some(*n),
             _ => None,
         }
     }
